Add netheriteAxeInstantMineWood rule
Mixin for AbstractClientPlayerEntity and ServerPlayerEntity added
which will override the getBlockBreakingSpeed method. The efficiency
levels and haste amplifier on the player are calculated and if they
satisfy the efficiency V haste II requirement will then check whether a
netherite axe is in the main hand. If this is also satisfied, and the
rule is enabled, then the netherite axe will break wooden material blocks
as if it was a golden axe with efficiency V and haste III.