Refactor instant mining rules to use composition
Logic for the instant mining rules with netherite has now been modified
to be more reusable and scalable using composition. The calculators for
individual tools are now enum values. New rules can be added more easily
this way. However, each rule addition needs a new key and accessor in
InstantMiningCarpetRuleKeys and InstantMiningCarpetRuleAccessor
respectively.

The wood rule has now been changed to use blocks instead of the material
so that it is consistent with the other rules. This also means more
control over which blocks are legible for each rule.