Refactor PhantomSpawner_PhantomsObeyHostileMobCapMixin to work per player
Introduced a new mixin: SpawnHelperInfoInvokerMixin
 - used to access the isBelowCap private method in SpawnHelper.Info

PhantomSpawner_PhantomsObeyHostileMobCapMixin has been refactored to
use this method, which reduces code duplication and will work per player.
The target method is now to redirect the isSpectator method of a player,
so when all world players are accessed in the phantomSpawner class, the
ability to spawn will be determined by whether the per player cap is full
or not.