Add phantomspawning package to carpetaddonsnotfound
New package which is concerned with phantom spawning in the game. We
have several rules pertaining to phantom spawning in the mod and this
package aims to decouple each rule implementation from one another. This
also aims to fix incompatability issues with other mods as most seem to
target the same method with a redirect in PhantomSpawner. Handlers are
introduced in this package which allow chaining of a "can a phantom
spawn" request. This will make the mixin extremely lightweight and easier
to adapt in the future if other incompatability issues are found.