Refactor ServerChunkManager_PhantomsObeyHostileMobCapMixin to avoid NullPointerException
ServerChunkManager_PhantomsObeyHostileMobCapMixin will now always set
the SpawnHelper.Info value even if the rule is not enabled. This is
required following the refactoring of the PhantomSpawner mixin. If we
did not do this, a NullPointerException is thrown in the
ObeyHostileMobCapHandler class. We could refactor
ObeyHostileMobCapHandler but it will lead to messier code.