Fix Stonecutter damage not producing damage sound
The `DamageType` `RegistryEntry` was a custom one but there is no way
to currently register custom DamageTypes in Fabric, as there is no
dedicated method in the `Registries` class for `DamageType`. The client
side resolves a packet which expects the damage type to be registered in
the registry. Removed the "stonecutter" custom one in favour of the
"GENERIC" version.