Add `netherWater` carpet rule (#232)
Adds the carpet rule `netherWater` which allows for water sources to be
generated in the nether when a charged creeper blows up an ice block.

Also fixes a bug with the custom carpet rule translations file reader
for 1.17.1 and 1.18.2 versions which wouldn't read the descriptions and
extras from the file. This doesn't affect any published version.

Closes #216